requirements.txt
flask
requests
python-dotenv
tldextract
validators
python-levenshtein
beautifulsoup4
